+++ dynamite/analysis.py
@@ -63,14 +63,14 @@ class Decomposition:
         self.logger.info(f'Performing decomposition for kin_set no {kin_set}: '
                          f'{stars.kinematic_data[kin_set].name}')
         # Get losvd_histograms and projected_masses
-        orblib = self.model.get_orblib()
-        orblib.read_losvd_histograms()
-        self.losvd_histograms = orblib.losvd_histograms[self.kin_set]
-        self.proj_mass = orblib.projected_masses[self.kin_set]
+        self.orblib = self.model.get_orblib()
+        self.orblib.read_losvd_histograms()
+        self.losvd_histograms = self.orblib.losvd_histograms[self.kin_set]
+        self.proj_mass = self.orblib.projected_masses[self.kin_set]
         self.logger.debug(f'{self.losvd_histograms.y.shape=}, '
                           f'{self.proj_mass.shape=}.')
         # Get orbit weights and store them in self.model.weights
-        _ = self.model.get_weights(orblib)
+        _ = self.model.get_weights(self.orblib)
         # Do the decomposition
         self.decomp = self.decompose_orbits()
         # self.losvd_histograms, self.proj_mass, self.decomp = self.run_dec()
@@ -204,10 +204,8 @@ class Decomposition:
         conversion_factor=self.config.all_models.system.distMPc*1.0e6*1.49598e8
 
         ncol = n_dither ** 3
-        orbclass1 = np.genfromtxt(file2).T
-        orbclass1 = orbclass1.reshape((5,ncol,n_orb), order='F')
-        orbclass2 = np.genfromtxt(file3).T
-        orbclass2 = orbclass1.reshape((5,ncol,n_orb), order='F')
+        orbclass1=self.orblib.read_orbit_property_file_base(file2, ncol, n_orb)
+        orbclass2=self.orblib.read_orbit_property_file_base(file3, ncol, n_orb)
 
         orbw = self.model.weights
         n_orbs = len(orbw)
@@ -329,22 +327,22 @@ class Decomposition:
         t = []
         totalf = 0
         for i in range(len(comps)):
-                labels = [comps[i] + qq for qq in quant]
-                flux = comp_kinem_moments[labels[0]]
-                w = weights[[comps[i] in s for s in self.decomp['component']]]
-                fhist, fbinedge = np.histogram(grid[s_wide], bins=len(flux))
-                flux = flux / fhist
-                tt = flux[grid]*1.
-                tt = tt * np.sum(w)/np.sum(tt)
-                t.append(tt.copy())
-                if comps[i] in ['thin_d', 'warm_d', 'bulge']:
-                    totalf += np.sum(tt)
-                    if comps[i] == 'thin_d':
-                        fluxtot = tt
-                    else:
-                        fluxtot += tt
-                vel.append(comp_kinem_moments[labels[1]])
-                sig.append(comp_kinem_moments[labels[2]])
+            labels = [comps[i] + qq for qq in quant]
+            flux = comp_kinem_moments[labels[0]]
+            w = weights[[comps[i] in s for s in self.decomp['component']]]
+            fhist, fbinedge = np.histogram(grid[s_wide], bins=len(flux))
+            flux = flux / fhist
+            tt = flux[grid]*1.
+            tt = tt * np.sum(w)/np.sum(tt)
+            t.append(tt.copy())
+            if comps[i] in ['thin_d', 'warm_d', 'bulge']:
+                totalf += np.sum(tt)
+                if comps[i] == 'thin_d':
+                    fluxtot = tt
+                else:
+                    fluxtot += tt
+            vel.append(comp_kinem_moments[labels[1]])
+            sig.append(comp_kinem_moments[labels[2]])
 
         t = t/totalf
 
@@ -361,10 +359,10 @@ class Decomposition:
 
         table = {'x/arcs':xi_t,'y/arcs':yi_t}
         for i in range(len(comps)):
-                labels = [comps[i] + qq for qq in quant]
-                table.update({labels[0]:t[i][s],
-                             labels[1]:vel[i][grid[s]],
-                             labels[2]:sig[i][grid[s]]})
+            labels = [comps[i] + qq for qq in quant]
+            table.update({labels[0]:t[i][s],
+                         labels[1]:vel[i][grid[s]],
+                         labels[2]:sig[i][grid[s]]})
         comps_kin = astropy.table.Table(table)
 
         kin_name = stars.kinematic_data[self.kin_set].name
@@ -455,7 +453,8 @@ class Analysis:
                                     model=None,
                                     kin_set=None,
                                     v_sigma_option='fit',
-                                    kinematics_as='table'):
+                                    kinematics_as='table',
+                                    weights=None):
         """
         Generates an astropy table in the model directory that holds the
         model's data for creating Gauss-Hermite kinematic maps:
@@ -480,6 +479,9 @@ class Analysis:
             format and return its full path ``f_name``, if 'both', write the
             table to disk and return a tuple ``(gh_table, f_name)``.
             The default is 'table'.
+        weights : ``numpy.array`` like, optional
+            Orbital weights to use. The default is ``None`` and will
+            determine the weights via ``model.get_weights(orblib)``.
 
         Raises
         ------
@@ -514,8 +516,9 @@ class Analysis:
         kin_name = stars.kinematic_data[kin_set].name
         self.logger.info('Getting model projected masses and losvds.')
         orblib = model.get_orblib()
-        _ = model.get_weights(orblib)
-        weights = model.weights
+        if weights is None:
+            _ = model.get_weights(orblib)
+            weights = model.weights
         # get losvd_histograms and projected masses:
         orblib.read_losvd_histograms()
         # get all orbits' losvds; orbits_losvd.shape = n_orb,n_vbin,n_aperture