+++ dynamite/data_prep/generate_kin_input.py
@@ -20,18 +20,18 @@ def read_kinematics_user(file):
 def read_atlas3d(file):
     #adapted from http://www-astro.physics.ox.ac.uk/atlas3d/
     #read original cube
-    
+
     hdu = fits.open(file[0])
     spectrum = hdu[0].data
     table = hdu[2].data
     #hdr = hdu[2].header #unused
-    
-    x = table["A"] # Coordinates of the original spaxels in arcsec 
+
+    x = table["A"] # Coordinates of the original spaxels in arcsec
     y = table["D"]
     flux = np.mean(spectrum, 1)  #surface brightness
-    
+
     hdu = fits.open(file[1])
-    table = hdu[1].data 
+    table = hdu[1].data
     #kin_hdr=hdu[1].header #unused
 
     xgen = table['XS'] # Voronoi generators
@@ -44,27 +44,36 @@ def read_atlas3d(file):
     dsig=table['ESPXF']
     dh3=table['EH3PXF']
     dh4=table['EH4PXF']
-    
+
     #perform Voronoi tesselation starting from the nodes values, adapted from axisymm Schwarzschild code
     npixels=len(x)
-    binNum=np.zeros((npixels), dtype=int)  
+    binNum=np.zeros((npixels), dtype=int)
     for j, (xj, yj) in enumerate(zip(x, y)):
         binNum[j] = np.argmin(((xj - xgen)**2 + (yj - ygen)**2))
 
     return binNum,x,y,flux,vel,sig,h3,h4,dvel,dsig,dh3,dh4,xgen,ygen
-    
+    #can be switched on for a check. They should show the same. If not, something went wrong with the V. tesselation
+    #plt.figure()
+    #im1=dbg.display_bins_generators(xgen,ygen,vel,x,y,label='Velocity (km/s)')
+    #plt.figure()
+    #im1=dp.display_pixels(x,y,vel[binNum],label='Velocity (km/s)')
+
 def read_califa(file):
     hdulist = fits.open(file)
     kin_tab = hdulist[1].data
     #kin_hdr = hdulist[1].header #unused
+
     s = kin_tab.BIN_ID > 0
+
     binNum = (kin_tab[s].BIN_ID).astype(int)
+
     xp = kin_tab[s].X
     yp = kin_tab[s].Y
     flux=kin_tab[s].FLUX
-    
+
     #check where each bin appears the first time
     ubins, indices = np.unique(binNum, return_index=True)
+
     vel = kin_tab[s].VPXF[indices]
     sig = kin_tab[s].SPXF[indices]
     h3 = kin_tab[s].H3PXF[indices]
@@ -82,7 +91,7 @@ def read_califa(file):
     mingherr = 0.005
     dh3[dh3 < mingherr] = mingherr
     dh4[dh4 < mingherr] = mingherr
-    
+
     # adopted from schwpy routine
     # Check that the dispersion is positive
     if np.min(sig) < 0:
@@ -90,6 +99,7 @@ def read_califa(file):
         mask = sig < 0
         sig[mask] = np.median(sig[sig >= 0])
         dsig[mask] = 1e5
+
     nbin = int(np.max(kin_tab.BIN_ID))
     xbin = np.zeros(nbin)
     ybin = np.zeros(nbin)
@@ -97,7 +107,7 @@ def read_califa(file):
         si = np.where(ubins == i+1)[0]
         xbin[i] = kin_tab[si].X
         ybin[i] = kin_tab[si].Y
-    
+
     return binNum-1,xp,yp,flux,vel,sig,h3,h4,dvel,dsig,dh3,dh4,xbin,ybin
 
 def create_aperture_file(dir,expr,minx,maxx,miny,maxy,angle_deg,nx,ny):
@@ -109,7 +119,6 @@ def create_aperture_file(dir,expr,minx,maxx,miny,maxy,angle_deg,nx,ny):
     aperture_file.write('\t{0:<.6f}\t{1:<.6f} \n'.format(maxx-minx, maxy-miny))
     aperture_file.write('\t{0:<.6f} \n'.format(angle_deg))
     aperture_file.write('\t{0}\t{1} \n'.format(int(nx), int(ny)))
-    aperture_file.write(' aperture = -(hst_pa) + 90 \n')
     aperture_file.close()
 
 def create_bins_file(dir,expr,grid):
@@ -121,8 +130,8 @@ def create_bins_file(dir,expr,grid):
     num_of_lines = int(len(flattened)/10)
     bins_file = open(dir+'bins'+expr+'.dat', 'w')
     bins_file.write('#Counterrotaton_binning_version_1\n')
-    bins_file.write('{0}\n'.format(int(s[0]*s[1])))  
-    
+    bins_file.write('{0}\n'.format(int(s[0]*s[1])))
+
     for line in range(num_of_lines):
         string = '\t{0}\t{1}\t{2}\t{3}\t{4}\t{5}\t{6}\t{7}\t{8}\t{9}\n'.format(int(flattened[line*10]),int(flattened[line*10+1]), 
                                                                                int(flattened[line*10+2]),int(flattened[line*10+3]), 
@@ -141,29 +150,30 @@ def create_bins_file(dir,expr,grid):
 
 def kin_file(dir,expr,data):
     data.write(dir+'gauss_hermite_kins'+expr +'.ecsv', format='ascii.ecsv', overwrite=True)
-    
-def create_kin_input(object, file, dyn_model_dir, expr='', angle_deg=0, ngh=4, 
+
+#### main routine ######
+def create_kin_input(object, file, dyn_model_dir, expr='', angle_deg=0, ngh=4,
                      pointsym=0, bisym=0, xoffset=0, yoffset=0, voffset=0, fit_PA=False, kin_input=0, files=True,plot=True):
-    
+
     """
     Create the inputs from the Voronoi binned kinematics files
     Returns bins.dat, aperture.dat and kin_data.dat and plots the kinematics
-    
+
     Note: sym keywords will be added later!
     """
     print('Galaxy: {0}'.format(object))
     print(file)
-    
+
     #read in binned kinematics, this needs to be changed by the user
     if kin_input == 'CALIFA':
         binNum,xp,yp,flux,vel,sig,h3,h4,dvel,dsig,dh3,dh4,xbin,ybin=read_califa(file)
-        
+
     elif kin_input == 'ATLAS3D':
         binNum,xp,yp,flux,vel,sig,h3,h4,dvel,dsig,dh3,dh4,xbin,ybin=read_atlas3d(file)
-        
+
     elif kin_input == 'USER':
         binNum,xp,yp,flux,vel,sig,h3,h4,dvel,dsig,dh3,dh4,xbin,ybin=read_kinematics_user(file)
-    
+
     nbins = len(vel)
 
     xp=xp+xoffset
@@ -179,8 +189,8 @@ def create_kin_input(object, file, dyn_model_dir, expr='', angle_deg=0, ngh=4,
         angle_deg,_,vel_syst = pa.fit_kinematic_pa(xp,yp,vel[binNum],cmap=map2) 
         plt.savefig(dyn_model_dir+'pafit.pdf')
         vel=vel - vel_syst
-    vel = vel+ voffset
-    
+    vel = vel + voffset
+
     # Determination of the pixel size
     npixels = len(xp)
     dx = 1e30
@@ -202,7 +212,7 @@ def create_kin_input(object, file, dyn_model_dir, expr='', angle_deg=0, ngh=4,
     nx = np.round((maxx-minx)/dx)
     ny = np.round((maxy-miny)/dx)
     print("Pixel grid dimension is ", nx, ny)
-    
+
     grid = np.zeros((int(nx), int(ny)))
     k = ((xp-minx)/dx).astype(int)
     j = ((yp-miny)/dx).astype(int)
@@ -214,27 +224,27 @@ def create_kin_input(object, file, dyn_model_dir, expr='', angle_deg=0, ngh=4,
         smax = np.percentile(sig, 98)
         smin = np.percentile(sig, 2)
         print('Vels plot: {0}, {1}, {2}'.format(vmax, smin, smax))
-        
+
         fig, axs = plt.subplots(1, 4, figsize=(18,4))
         plt.subplot(1,4,1)
         plt.title('Velocity [km/s]')
         dp.display_pixels(xp,yp,vel[binNum],angle=angle_deg,vmin=-vmax,vmax=vmax,cmap=map2)
-        
+
         plt.subplot(1,4,2)
         plt.title('Velocity dispersion [km/s]')
         dp.display_pixels(xp,yp,sig[binNum],angle=angle_deg,vmin=smin,vmax=smax,cmap=map1)
-        
+
         plt.subplot(1,4,3)
         plt.title(r'$h_{3}$ moment')
         dp.display_pixels(xp,yp,h3[binNum],angle=angle_deg,vmin=-0.15,vmax=0.15,cmap=map2)
-        
+
         plt.subplot(1,4,4)
         plt.title(r'$h_{4}$ moment')
         dp.display_pixels(xp,yp,h4[binNum],angle=angle_deg,vmin=-0.15,vmax=0.15,cmap=map2)
-        
+
         fig.subplots_adjust(left=0.04, wspace=0.3, hspace=0.01, right=0.97)
         fig.savefig(dyn_model_dir+'kinmaps.pdf')
-    
+
     data = table.Table()
     data['vbin_id']=np.arange(1,nbins+1)
     data['v']=np.round(vel,decimals=4)
@@ -244,8 +254,8 @@ def create_kin_input(object, file, dyn_model_dir, expr='', angle_deg=0, ngh=4,
     data['h3']=np.round(h3,decimals=4)
     data['dh3']=dh3
     data['h4']=np.round(h4,decimals=4)
-    data['dh4']=dh4  
-    
+    data['dh4']=dh4
+
     if ngh==6:
         if (kin_input=='CALIFA') or ((kin_input=='ATLAS3D')):
             h5=np.full_like(h4, 0)
@@ -255,8 +265,8 @@ def create_kin_input(object, file, dyn_model_dir, expr='', angle_deg=0, ngh=4,
         data['h5']=np.round(h5,decimals=4)
         data['dh5']=dh5
         data['h6']=np.round(h6,decimals=4)
-        data['dh6']=dh6  
-    
+        data['dh6']=dh6
+
     if files is True:
         create_aperture_file(dyn_model_dir,expr,minx,maxx,miny,maxy,angle_deg,nx,ny)
         create_bins_file(dyn_model_dir,expr,grid)