+++ docs/more_info/changelog.rst
@@ -5,6 +5,12 @@ Change Log
 ****************
 
 - Improvement: updated tutorial notebooks
+
+- New feature: kinmapchi2 (directly calculated from the kinematic maps) is now also available for the python NNLS solver
+- New feature: added support for bar/disk decomposition
+- New feature: added support for getting intrinsic model moments for both Gauss Hermite and a BayesLOSVD models
+- Improvement: Eliminated unused position_angle system attribute from the configuration file (the angle is read from aperture.dat)
+- Improvement: DYNAMITE can now be built without GALAHAD (LegacyWeightSolver will not be available then)
 - Improvement: plotting gh kinematic maps is more efficient and now works for all weight solvers
 - New feature: New parameter generator SpecificModels generates and runs a predefined list of models or models resulting from a cartesian product of parameter values
 - Improvement: the orbit plot (Plotter.orbit_plot) now works for all implemented weight solvers
@@ -16,7 +22,7 @@ Change Log
 - New feature: Added a new method AllModels.make_best_models_table() that creates a table of the best models (best n models or models within a chi2-threshold of the best) and saves it to disk
 - Bugfix: If reattempt_failures=False, in certain cases it could occur that orblibs of successful models were deleted
 - Bugfix: Fixed a bug related to a nonexistent model directory if a crash occurs between the parameter generator adding a model and starting to solve it
-- Improvement: Dynamite will no longer crash upon Legacy Fortran errors, but issue warnings and assign nan to the affected chi2 values
+- Improvement: Dynamite will no longer crash upon Legacy Fortran errors (except when executables are not found), but issue warnings and assign nan to the affected chi2 values
 - Improvement: When executing a dummy run (do_dummy_run==True), model_iterator will set both kinchi2 and kinmapchi2 to nan (instead of zero)
 - Improvement: DYNAMITE will retrofit existing all_models tables with the new column kinmapchi2 and calculate its values for existing models whenever possible
 - New feature: chi2 can now be directly calculated from the kinematic maps when using the LegacyWeightSolver via which_chi2: "kinmapchi2"